track the names of translators on translator instances.
translators are registered with an extension using a unique name for the translator to distinguish
it from alternate translators for the same node. which this change, the name assigned to a
translator class is attached to each instance (keep in mind, a single translator class can
be associated with multiple maya nodes, and thus multiple instances may receive different translator names).