Refactor COptionsTranslator file naming code, adding support for multi-channel exrs. (Ticket #192)
All code related to image naming is moved out of CAOV and into COptionsTranslator.
Adds a new splitAOVs attribute to aiArnoldDriver. when enabled (the default), this ensures that a <RenderPass> token is added to all image file paths.
When disabled, the prefix paths are used AS-IS, and multiple outputs with the same name will be merged (if the driver supports it).